// lib/generated/models/Coupon.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

export type Coupon = {
  /**
   * @type integer
   */
  readonly id: number
  /**
   * @maxLength 50
   * @type string
   */
  code: string
  /**
   * @type string
   */
  description: string
  /**
   * @minLength 0
   * @maxLength 2147483647
   * @type integer
   */
  max_redemptions?: number | null
  /**
   * @type boolean | undefined
   */
  available?: boolean
  /**
   * @type string, date-time
   */
  readonly created_at: string
}